+++ taucoin-android/src/main/java/io/taucoin/torrent/publishing/core/storage/sqlite/TxRepository.java
@@ -29,7 +29,7 @@ public interface TxRepository {
      */
     List<UserAndTx> getTxsByChainID(String chainID);
 
-    DataSource.Factory<Integer, UserAndTx> queryCommunityTxs(String chainID, int txType, int txStatus);
+    DataSource.Factory<Integer, UserAndTx> queryCommunityTxs(String chainID, long txType, int txStatus);
 
     /**
      * 获取社区里用户未上链并且未过期的交易数
